UCUnits::resolveResource: Prefer image path if it exists
This saves searching the disk for @gu images, wWhich is a big speedup

According to callgrind loading 100 images from a folder that contains 380 images goes from
around 3 million instructions per UCUnits::resolveResource call down to around 10 thousand
            
This optimization is ok since it is not correct to ship both
      image.png
and
      image@20.png
you either have to ship gu-enabled files or not, but mixing them is not allowed.
            
Searched for cases in which that may be happening in my phone and found none.